+++ Sources/BraintreePayPal/BTPayPalCheckoutRequest.swift
@@ -58,53 +58,41 @@ import BraintreeCore
 /// Options for the PayPal Checkout flow.
 @objcMembers open class BTPayPalCheckoutRequest: BTPayPalRequest {
 
-    // MARK: - Public Properties
-
-    ///  Used for a one-time payment.
-    ///
-    ///  Amount must be greater than or equal to zero, may optionally contain exactly 2 decimal places separated by '.' and is limited to 7 digits before the decimal point.
-    public var amount: String
-
-    /// Optional: Payment intent. Defaults to `.authorize`. Only applies to PayPal Checkout.
-    public var intent: BTPayPalRequestIntent
-
-    /// Optional: Changes the call-to-action in the PayPal Checkout flow. Defaults to `.none`.
-    public var userAction: BTPayPalRequestUserAction
-
-    /// Optional: Offers PayPal Pay Later if the customer qualifies. Defaults to `false`. Only available with PayPal Checkout.
-    public var offerPayLater: Bool
-
-    /// Optional: A three-character ISO-4217 ISO currency code to use for the transaction. Defaults to merchant currency code if not set.
-    ///
-    /// - Note: See https://developer.paypal.com/docs/api/reference/currency-codes/ for a list of supported currency codes.
-    public var currencyCode: String?
-
-    /// Optional: If set to `true`, this enables the Checkout with Vault flow, where the customer will be prompted to consent to a billing agreement during checkout. Defaults to `false`.
-    public var requestBillingAgreement: Bool
+    // MARK: - Internal Properties
     
-    /// Optional: User email to initiate a quicker authentication flow in cases where the user has a PayPal Account with the same email.
-    public var userAuthenticationEmail: String?
+    var amount: String
+    var intent: BTPayPalRequestIntent
+    var userAction: BTPayPalRequestUserAction
+    var offerPayLater: Bool
+    var currencyCode: String?
+    var requestBillingAgreement: Bool
 
     // MARK: - Initializer
 
     /// Initializes a PayPal Native Checkout request
     /// - Parameters:
-    ///   - amount: Used for a one-time payment. Amount must be greater than or equal to zero, may optionally contain exactly 2 decimal places separated by '.'
+    ///   - amount: Used for a one-time payment. Amount must be greater than or equal to zero, may optionally contain exactly 2 decimal places separated by '.' and is limited to 7 digits before the decimal point.
     ///   - intent: Optional: Payment intent. Defaults to `.authorize`. Only applies to PayPal Checkout.
-    ///   and is limited to 7 digits before the decimal point.
     ///   - userAction: Optional: Changes the call-to-action in the PayPal Checkout flow. Defaults to `.none`.
     ///   - offerPayLater: Optional: Offers PayPal Pay Later if the customer qualifies. Defaults to `false`. Only available with PayPal Checkout.
     ///   - currencyCode: Optional: A three-character ISO-4217 ISO currency code to use for the transaction. Defaults to merchant currency code if not set.
     ///   See https://developer.paypal.com/docs/api/reference/currency-codes/ for a list of supported currency codes.
     ///   - requestBillingAgreement: Optional: If set to `true`, this enables the Checkout with Vault flow, where the customer will be prompted to consent to a billing agreement
     ///   during checkout. Defaults to `false`.
+    ///   - userAuthenticationEmail: Optional: User email to initiate a quicker authentication flow in cases where the user has a PayPal Account with the same email.
+    ///   - userPhoneNumber: Optional: A user's phone number to initiate a quicker authentication flow in the scenario where the user has a PayPal account
+    ///   identified with the same phone number.
+    ///   - lineItems: Optional: The line items for this transaction. It can include up to 249 line items.
     public init(
         amount: String,
         intent: BTPayPalRequestIntent = .authorize,
         userAction: BTPayPalRequestUserAction = .none,
         offerPayLater: Bool = false,
         currencyCode: String? = nil,
-        requestBillingAgreement: Bool = false
+        requestBillingAgreement: Bool = false,
+        userAuthenticationEmail: String? = nil,
+        userPhoneNumber: BTPayPalPhoneNumber? = nil,
+        lineItems: [BTPayPalLineItem]? = nil
     ) {
         self.amount = amount
         self.intent = intent
@@ -112,8 +100,13 @@ import BraintreeCore
         self.offerPayLater = offerPayLater
         self.currencyCode = currencyCode
         self.requestBillingAgreement = requestBillingAgreement
-
-        super.init(hermesPath: "v1/paypal_hermes/create_payment_resource", paymentType: .checkout)
+        super.init(
+            hermesPath: "v1/paypal_hermes/create_payment_resource",
+            paymentType: .checkout,
+            lineItems: lineItems,
+            userAuthenticationEmail: userAuthenticationEmail,
+            userPhoneNumber: userPhoneNumber
+        )
     }
 
     // MARK: Internal Methods
@@ -134,10 +127,6 @@ import BraintreeCore
 
         if currencyCode != nil {
             checkoutParameters["currency_iso_code"] = currencyCode
-        }
-        
-        if let userAuthenticationEmail, !userAuthenticationEmail.isEmpty {
-            checkoutParameters["payer_email"] = userAuthenticationEmail
         }
 
         if userAction != .none, var experienceProfile = baseParameters["experience_profile"] as? [String: Any] {