+++ Sources/BraintreePayPal/BTPayPalRequest.swift
@@ -52,52 +52,22 @@ import BraintreeCore
 /// - Note: Do not instantiate this class directly. Instead, use BTPayPalCheckoutRequest or BTPayPalVaultRequest.
 @objcMembers open class BTPayPalRequest: NSObject {
 
-    // MARK: - Public Properties
-
-    /// Defaults to false. When set to true, the shipping address selector will be displayed.
-    public var isShippingAddressRequired: Bool
-
-    /// Defaults to false. Set to true to enable user editing of the shipping address.
-    /// - Note: Only applies when `shippingAddressOverride` is set.
-    public var isShippingAddressEditable: Bool
-
-    ///  Optional: A locale code to use for the transaction.
-    public var localeCode: BTPayPalLocaleCode
-
-    /// Optional: A valid shipping address to be displayed in the transaction flow. An error will occur if this address is not valid.
-    public var shippingAddressOverride: BTPostalAddress?
-
-    /// Optional: Landing page type. Defaults to `.none`.
-    /// - Note: Setting the BTPayPalRequest's landingPageType changes the PayPal page to display when a user lands on the PayPal site to complete the payment.
-    ///  `.login` specifies a PayPal account login page is used.
-    ///  `.billing` specifies a non-PayPal account landing page is used.
-    public var landingPageType: BTPayPalRequestLandingPageType
-
-    /// Optional: The merchant name displayed inside of the PayPal flow; defaults to the company name on your Braintree account
-    public var displayName: String?
-
-    /// Optional: A non-default merchant account to use for tokenization.
-    public var merchantAccountID: String?
-
-    /// Optional: The line items for this transaction. It can include up to 249 line items.
-    public var lineItems: [BTPayPalLineItem]?
-
-    /// Optional: Display a custom description to the user for a billing agreement. For Checkout with Vault flows, you must also set
-    ///  `requestBillingAgreement` to `true` on your `BTPayPalCheckoutRequest`.
-    public var billingAgreementDescription: String?
-
-    /// Optional: A risk correlation ID created with Set Transaction Context on your server.
-    public var riskCorrelationID: String?
-    
-    /// Optional: A user's phone number to initiate a quicker authentication flow in the scenario where the user has a PayPal account
-    /// identified with the same phone number.
-    public var userPhoneNumber: BTPayPalPhoneNumber?
-
     // MARK: - Internal Properties
     
     var hermesPath: String
-
     var paymentType: BTPayPalPaymentType
+    var isShippingAddressRequired: Bool
+    var isShippingAddressEditable: Bool
+    var localeCode: BTPayPalLocaleCode?
+    var shippingAddressOverride: BTPostalAddress?
+    var landingPageType: BTPayPalRequestLandingPageType?
+    var displayName: String?
+    var merchantAccountID: String?
+    var lineItems: [BTPayPalLineItem]?
+    var billingAgreementDescription: String?
+    var riskCorrelationID: String?
+    var userAuthenticationEmail: String?
+    var userPhoneNumber: BTPayPalPhoneNumber?
     
     // MARK: - Static Properties
     
@@ -105,6 +75,27 @@ import BraintreeCore
 
     // MARK: - Initializer
 
+    /// Initialize a new `BTPayPalRequest`
+    /// - Parameters:
+    ///    - hermesPath: Required :nodoc: The hermes path or endpoint URI path. This property is not covered by semantic versioning.
+    ///    - paymentType: Required :nodoc: The payment type, either checkout or vault. This property is not covered by semantic versioning.
+    ///    - isShippingAddressRequired: Defaults to false. When set to true, the shipping address selector will be displayed.
+    ///    - isShippingAddressEditable: Defaults to false. Set to true to enable user editing of the shipping address. 
+    ///     - Note: Only applies when `shippingAddressOverride` is set.
+    ///    - localeCode: Optional: A locale code to use for the transaction.
+    ///    - shippingAddressOverride: Optional: A valid shipping address to be displayed in the transaction flow. An error will occur if this address is not valid.
+    ///    - landingPageType: Optional: Landing page type. Defaults to `.none`.  
+    ///     - Note: Setting the BTPayPalRequest's landingPageType changes the PayPal page to display when a user lands on the PayPal site to complete the payment.
+    ///        `.login` specifies a PayPal account login page is used.
+    ///        `.billing` specifies a non-PayPal account landing page is used.
+    ///    - displayName: Optional: The merchant name displayed inside of the PayPal flow; defaults to the company name on your Braintree account
+    ///    - merchantAccountID: Optional: A non-default merchant account to use for tokenization.
+    ///    - lineItems: Optional: The line items for this transaction. It can include up to 249 line items.
+    ///    - billingAgreementDescription: Optional: Display a custom description to the user for a billing agreement. For Checkout with Vault flows, you must also set
+    ///        `requestBillingAgreement` to `true` on your `BTPayPalCheckoutRequest`.
+    ///    - riskCorrelationID: Optional: A risk correlation ID created with Set Transaction Context on your server.
+    ///    - userAuthenticationEmail: Optional: User email to initiate a quicker authentication flow in cases where the user has a PayPal Account with the same email.
+    ///    - userPhoneNumber: Optional: A user's phone number to initiate a quicker authentication flow in the scenario where the user has a PayPal account identified with the same phone number.
     init(
         hermesPath: String,
         paymentType: BTPayPalPaymentType,
@@ -117,7 +108,8 @@ import BraintreeCore
         merchantAccountID: String? = nil,
         lineItems: [BTPayPalLineItem]? = nil,
         billingAgreementDescription: String? = nil,
-        riskCorrelationId: String? = nil,
+        riskCorrelationID: String? = nil,
+        userAuthenticationEmail: String? = nil,
         userPhoneNumber: BTPayPalPhoneNumber? = nil
     ) {
         self.hermesPath = hermesPath
@@ -131,7 +123,8 @@ import BraintreeCore
         self.merchantAccountID = merchantAccountID
         self.lineItems = lineItems
         self.billingAgreementDescription = billingAgreementDescription
-        self.riskCorrelationID = riskCorrelationId
+        self.riskCorrelationID = riskCorrelationID
+        self.userAuthenticationEmail = userAuthenticationEmail
         self.userPhoneNumber = userPhoneNumber
     }
 
@@ -147,12 +140,12 @@ import BraintreeCore
         experienceProfile["no_shipping"] = !isShippingAddressRequired
         experienceProfile["brand_name"] = displayName != nil ? displayName : configuration.json?["paypal"]["displayName"].asString()
 
-        if landingPageType.stringValue != nil {
-            experienceProfile["landing_page_type"] = landingPageType.stringValue
+        if landingPageType?.stringValue != nil {
+            experienceProfile["landing_page_type"] = landingPageType?.stringValue
         }
 
-        if localeCode.stringValue != nil {
-            experienceProfile["locale_code"] = localeCode.stringValue
+        if localeCode?.stringValue != nil {
+            experienceProfile["locale_code"] = localeCode?.stringValue
         }
 
         experienceProfile["address_override"] = shippingAddressOverride != nil ? !isShippingAddressEditable : false
@@ -166,10 +159,14 @@ import BraintreeCore
         if riskCorrelationID != nil {
             parameters["correlation_id"] = riskCorrelationID
         }
-
+        
         if let lineItems, !lineItems.isEmpty {
             let lineItemsArray = lineItems.compactMap { $0.requestParameters() }
             parameters["line_items"] = lineItemsArray
+        }
+        
+        if let userAuthenticationEmail, !userAuthenticationEmail.isEmpty {
+            parameters["payer_email"] = userAuthenticationEmail
         }
         
         if let userPhoneNumberDict = try? userPhoneNumber?.toDictionary() {